Add scripts to create etcdClient permissions
Pks-nsx-t-ncp job has a post-start script which uses kubectl.
For a 1.7 => 1.8 upgrade, the other master instances will not have been upgraded yet.
Kubectl will intermittently fail as it round robins to other kube-apiservers

PKS 1.8 uses a new version of etcd which sets up authentication in the
v3 namespace, which is what triggers this behavior

[#173115867]

Co-authored-by: Larry Hamel <lhamel@pivotal.io>
Co-authored-by: Bridget Arthur <barthur@pivotal.io>
Co-authored-by: Daniel Keeney <dkeeney@pivotal.io>